+++ loofi-fedora-tweaks/ui/storage_tab.py
@@ -65,6 +65,7 @@ class StorageTab(BaseTab):
         self.disk_table.setEditTriggers(QTableWidget.EditTrigger.NoEditTriggers)
         self.disk_table.setSelectionBehavior(QTableWidget.SelectionBehavior.SelectRows)
         self.disk_table.setMaximumHeight(140)
+        BaseTab.configure_table(self.disk_table)
         dl_layout.addWidget(self.disk_table)
 
         disk_btn_layout = QHBoxLayout()
@@ -94,6 +95,7 @@ class StorageTab(BaseTab):
             QHeaderView.ResizeMode.Stretch
         )
         self.mount_table.setEditTriggers(QTableWidget.EditTrigger.NoEditTriggers)
+        BaseTab.configure_table(self.mount_table)
         ml_layout.addWidget(self.mount_table)
 
         layout.addWidget(mount_group)